+++ blocklets/inscription/api/libs/request.js
@@ -1,5 +1,8 @@
 const axios = require('axios');
 
+const api = axios.create({});
+
 module.exports = {
   axios,
+  api,
 };